Fix bustage from bug 536324 - Change nsIChannel's contentLength to be 64-bit. r=bustage-fix CLOSED TREE

+++ mailnews/base/util/nsMsgProtocol.cpp
@@ -678,13 +678,13 @@ nsMsgProtocol::GetContentDispositionHeader(nsACString &aContentDispositionHeader
   return NS_ERROR_NOT_AVAILABLE;
 }
 
-NS_IMETHODIMP nsMsgProtocol::GetContentLength(int32_t *aContentLength)
+NS_IMETHODIMP nsMsgProtocol::GetContentLength(int64_t *aContentLength)
 {
   *aContentLength = mContentLength;
   return NS_OK;
 }
 
-NS_IMETHODIMP nsMsgProtocol::SetContentLength(int32_t aContentLength)
+NS_IMETHODIMP nsMsgProtocol::SetContentLength(int64_t aContentLength)
 {
   mContentLength = aContentLength;
   return NS_OK;
