+++ mailnews/news/src/nsNntpMockChannel.cpp
@@ -232,7 +232,7 @@ nsNntpMockChannel::GetContentDispositionHeader(nsACString &aContentDispositionHe
   return NS_ERROR_NOT_AVAILABLE;
 }
 
-NS_IMETHODIMP nsNntpMockChannel::GetContentLength(int32_t *length)
+NS_IMETHODIMP nsNntpMockChannel::GetContentLength(int64_t *length)
 {
   FORWARD_CALL(GetContentLength, length)
   *length = m_contentLength;
@@ -240,7 +240,7 @@ NS_IMETHODIMP nsNntpMockChannel::GetContentLength(int32_t *length)
 }
 
 NS_IMETHODIMP
-nsNntpMockChannel::SetContentLength(int32_t aLength)
+nsNntpMockChannel::SetContentLength(int64_t aLength)
 {
   FORWARD_CALL(SetContentLength, aLength)
   m_contentLength = aLength;
